modules/locations-producer/requirements.txt
grpcio==1.39.0
protobuf==3.17.3
kafka-python==2.0.2
